add monitor state

+++ src/main/java/com/obernal/portal_monitoratge/model/monitor/MonitorMetadata.java
@@ -16,9 +16,9 @@ public class MonitorMetadata {
     private Set<String> labels;
     private String documentation;
     private boolean active;
-    private int insistAfter;
+    private long insistAfter;
 
-    public MonitorMetadata(String name, String description, String cron, String service, Set<String> labels, String documentation, int insistAfter) {
+    public MonitorMetadata(String name, String description, String cron, String service, Set<String> labels, String documentation, long insistAfter) {
         this.id = UUID.randomUUID().toString();
         created = LocalDateTime.now();
         updated = null;
@@ -32,7 +32,7 @@ public class MonitorMetadata {
         this.insistAfter = insistAfter;
     }
 
-    public MonitorMetadata(String id, LocalDateTime created, LocalDateTime updated, String name, String description, String cron, String service, Set<String> labels, String documentation, boolean active, int insistAfter) {
+    public MonitorMetadata(String id, LocalDateTime created, LocalDateTime updated, String name, String description, String cron, String service, Set<String> labels, String documentation, boolean active, long insistAfter) {
         this.id = id;
         this.created = created;
         this.updated = updated;
@@ -102,7 +102,7 @@ public class MonitorMetadata {
         return active;
     }
 
-    public int getInsistAfter() {
+    public long getInsistAfter() {
         return insistAfter;
     }
 }
